Lot of changes to get it to compile under string null checking.
Changed interface to vms so that getting values associated with paths can not return nulls. You must check whether the path is mapped.

Similarly, changed the interface to the VarStack so that getting an item from the var stack does not return nulls.  You must check that the field exists.

Changed the implementation of VarStack to use separate classes for empty and nonempty var stacks.