Changed the interpretation of moveRight so that when the selection is not empty, it goes down to the first position under the first selected node.  It's essentially a left to right pre-order traversal.
Changed the interpretation of moveLeft so that it is the inverse of moveRight, when the original selection is empty or size 1.  I.e. it is essentially a right to left post-order traversal. (Or the reverse of a left to right postorder traversal.

In a picture we have, for example (! is anchor and ^ is focus)
From top to bottom are successive applications of moveRight
From bottom to top are successive applications of moveLeft
   a( !^ b(), c(d()), f())
   a( !b()^, c(d()), f())
   a( b(!^), c(d()), f())
   a( b(), !^ c(d()), f())
   a( b(), !c(d())^, f())
   a( b(), c(!^d()), f())
   a( b(), c(!d()^), f())
   a( b(), c(d(!^)), f())
   a( b(), c(d()!^), f())
   a( b(), c(d()),!^ f())
   a( b(), c(d()), !f()^)
   a( b(), c(d()), f(!^))
   a( b(), c(d()), f()!^)

As an example of selection of size greater than 1 we have
     a( b(), ^c(d()), f()! )   ---moveRight---> a( b(), c( !^ d()), f())
and
     a( b(), ^c(d()), f()! )   ---moveLeft---> a( b(), !^ c(d()), f())